Create simple help service and directive
This commit adds 2 components:

1) A Help Service (gaHelpRegistry). This service can be used
to get a html snippet corresponding to a given help id. Right
now, this html snippets are hard coded, but in the future, they
will come either from some template, url or service. It's a
requirement that the help texts will be used outside of the
application. This needs to be defined first.

2) A Help Directive (gaHelp). This directive can be used to
integrate a help button anywhere in the page (or partial) and
associate it with a given help-id (configured
with ga-help attribute). When the button is clicked, the
help is displayed.

Help in popups can also be used using the new 'help'
options paramenter (with help id as value). This will display
a help button in any popup.

Help button is shown in the following components:
- Search (on hover only)
- Catalog/Topic in Accordeon (on hover only)
- Layer Manager in Accordeon (on hover only)
- WMS Import (in popup)
- KML Import (in popup)

This PR also adds tooltip infos for all tools.