Adding language change broadcast
The Language directive now broadcasts a gaLanguageChange message
that other controllers/directives can listen to. This allows
directive that rely on outside source which are language
dependant to reload their resources. This is needed because
our services deliver translated things (like catalog entries,
layer names, etc) which are not translated on the client.

The only directive now depending on this is the
backgroundlayerselector. Before this, the language in the combobox
didn't change. Now it does.