Lang falls back to default while it shouldn't
If the translation directive got the gaTopicChange event before the translations (e.g. fr.json) were received then we fell back to the default language. This was wrong. This commit fixes it by using scope.lang instead of $translate.uses() as the current lang when verifying whether the topic supports the current lang.