Make identify request for each layer
This is needed because each layer could have different timeInstant
parameter. The overall behaviour of the tooltip component
does not change.

Now, timeenabled layers are supported, taking into account current
state of timeslider (enabled/disabled, active/inactive on specific
year or not).

This asssures that only the desired tooltips are shown.