Add languages to global options
TranslationService will use the list of languages if the topics are not
leaded yet. So we are sure to use the fallback language if the lang in
permalink or most likely the language we get from the web browser is not
supported. Without this, the loading of layerConfig.json migth fail and
thus the portail will be unsuable.

This assumes that the default topic is available for the fallback
language.